feat(cart-page): lift data fetching and state to CartPage
